requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-filter==22.1
django-import-export==3.0.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
pytz==2022.7
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
